API updates

+++ app/schemas/walletTransactionResponse.py
@@ -1,13 +1,14 @@
 from pydantic import BaseModel, ConfigDict
 from typing import Optional
 from datetime import datetime
+from decimal import Decimal
 
 class WalletTransactionResponse(BaseModel):
     message: str
-    amount: float
+    amount: Decimal
     currency: str
     description: Optional[str] = None
-    balance: float
+    balance: Decimal
     transaction_type: str  # "deposit", "withdraw", "transfer"
     timestamp: datetime
     model_config=ConfigDict(from_attributes = True)
